Simplify now that #170 is fixed
Previously we had to explicitly refer to [_k - 1] in
the use of inductive predicates which were in the same
connected component as the "ensures" but not literally
mentioned in the "ensures". Now, Dafny correctly rewrites
inductive lemmas to insert decremented k-arguments for all
inductive predicates in the same connected component as
any mentioned in the "ensures". This simplifies things.